+++ Builds/VisualStudio2013/RippleD.vcxproj
@@ -3425,6 +3425,10 @@
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='debug|x64'">True</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='release|x64'">True</ExcludedFromBuild>
     </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\utility.cpp">
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='debug|x64'">True</ExcludedFromBuild>
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='release|x64'">True</ExcludedFromBuild>
+    </ClCompile>
     <ClInclude Include="..\..\src\ripple\test\jtx\json.h">
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\JTx.h">
@@ -3448,6 +3452,8 @@
     <ClInclude Include="..\..\src\ripple\test\jtx\ticket.h">
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\txflags.h">
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\utility.h">
     </ClInclude>
     <ClCompile Include="..\..\src\ripple\unity\app_ledger.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='debug.classic|x64'">True</ExcludedFromBuild>
