+++ Builds/VisualStudio2013/RippleD.vcxproj.filters
@@ -4149,6 +4149,9 @@
     <ClCompile Include="..\..\src\ripple\test\jtx\impl\txflags.cpp">
       <Filter>ripple\test\jtx\impl</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\ripple\test\jtx\impl\utility.cpp">
+      <Filter>ripple\test\jtx\impl</Filter>
+    </ClCompile>
     <ClInclude Include="..\..\src\ripple\test\jtx\json.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
@@ -4183,6 +4186,9 @@
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\ripple\test\jtx\txflags.h">
+      <Filter>ripple\test\jtx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\ripple\test\jtx\utility.h">
       <Filter>ripple\test\jtx</Filter>
     </ClInclude>
     <ClCompile Include="..\..\src\ripple\unity\app_ledger.cpp">
