+++ src/ripple/test/jtx/utility.h
@@ -17,30 +17,57 @@
 */
 //==============================================================================
 
-#ifndef RIPPLE_TEST_JTX_H_INCLUDED
-#define RIPPLE_TEST_JTX_H_INCLUDED
-
-// Convenience header that includes everything
+#ifndef RIPPLE_TEST_JTX_UTILITY_H_INCLUDED
+#define RIPPLE_TEST_JTX_UTILITY_H_INCLUDED
 
 #include <ripple/test/jtx/Account.h>
-#include <ripple/test/jtx/amounts.h>
-#include <ripple/test/jtx/any.h>
-//#include <ripple/test/jtx/balance.h>
-#include <ripple/test/jtx/Env.h>
-#include <ripple/test/jtx/fee.h>
-#include <ripple/test/jtx/json.h>
-#include <ripple/test/jtx/JTx.h>
-#include <ripple/test/jtx/multisign.h>
-#include <ripple/test/jtx/owners.h>
-#include <ripple/test/jtx/paths.h>
-#include <ripple/test/jtx/sendmax.h>
-#include <ripple/test/jtx/seq.h>
-#include <ripple/test/jtx/sig.h>
-#include <ripple/test/jtx/tags.h>
-#include <ripple/test/jtx/ter.h>
-#include <ripple/test/jtx/ticket.h>
-#include <ripple/test/jtx/txflags.h>
-#include <ripple/test/jtx/utility.h>
+#include <ripple/json/json_value.h>
+#include <ripple/app/ledger/Ledger.h>
+#include <ripple/protocol/STObject.h>
+#include <stdexcept>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+/** Set the fee automatically. */
+void
+fill_fee (Json::Value& jv,
+    Ledger const& ledger);
+
+/** Set the sequence number automatically. */
+void
+fill_seq (Json::Value& jv,
+    Ledger const& ledger);
+
+/** Sign automatically.
+    @note This only works on accounts with multi-signing off.
+*/
+void
+sign (Json::Value& jv,
+    Account const& account);
+
+/** Thrown when parse fails. */
+struct parse_error : std::logic_error
+{
+    template <class String>
+    explicit
+    parse_error (String const& s)
+        : logic_error(s)
+    {
+    }
+};
+
+/** Convert JSON to STObject.
+    This throws on failure, the JSON must be correct.
+    @note Testing malformed JSON is beyond the scope of
+          this set of unit test routines.
+*/
+STObject
+parse (Json::Value const& jv);
+
+} // jtx
+} // test
+} // ripple
 
 #endif
-
